fix: Add missing semicolon in loop and increment `it`
By default, `it` is not incremented because this line:

```cpp
it = ships.erase(it);
```

which automatically gets the next item in `std::list`. In those cases,
`it` should NOT be incremented.

Because of this, an `else` clause is needed where `it` IS incremented.
Otherwise, the `for` loop would loop forever because it stays at the
same item.

Signed-off-by: Kevin D <kevinduong306@gmail.com>